Use generics in concgroup
This updates the concgroup package to use generics instead of relying
on `interface{}`. Assuming that the use-case for concgroup is to always
return the same value type from `concgroup.Go` this should reduce some
boilerplate when using the library since we can avoid having to cast the
`interface{}`` values to the expected type.

This also adds two tests for the concgroup package to validate the
change and ensure that the concurrency limit is respected.